Make BlockByHash use async/await
Remove the old futures and make `BlockByHash` use async/await.  Note,
this patch removes the `impl_block_by_hash` macro in favour of
implementing this trait for both Ethereum and Bitcoin.  The macro relies
on a correct import in the file it is used in, this is racey.
