+++ cnd/src/btsieve/bitcoin/bitcoind_connector.rs
@@ -3,8 +3,6 @@ use crate::btsieve::{
 };
 use async_trait::async_trait;
 use bitcoin::{BlockHash, Network};
-use futures::Future;
-use futures_core::{compat::Future01CompatExt, FutureExt, TryFutureExt};
 use reqwest::{Client, Url};
 use serde::Deserialize;
 
@@ -43,9 +41,9 @@ impl LatestBlock for BitcoindConnector {
 
     async fn latest_block(&mut self) -> anyhow::Result<Self::Block> {
         let chaininfo_url = self.chaininfo_url.clone();
-        let this = self.clone();
+        let mut connector = self.clone();
 
-        let chain_info = this
+        let chain_info = connector
             .client
             .get(chaininfo_url)
             .send()
@@ -53,41 +51,29 @@ impl LatestBlock for BitcoindConnector {
             .json::<ChainInfo>()
             .await?;
 
-        let block = this
-            .block_by_hash(chain_info.bestblockhash)
-            .compat()
-            .await?;
+        let block = connector.block_by_hash(chain_info.bestblockhash).await?;
 
         Ok(block)
     }
 }
 
+#[async_trait]
 impl BlockByHash for BitcoindConnector {
     type Block = bitcoin::Block;
     type BlockHash = bitcoin::BlockHash;
 
-    fn block_by_hash(
-        &self,
-        block_hash: Self::BlockHash,
-    ) -> Box<dyn Future<Item = Self::Block, Error = anyhow::Error> + Send + 'static> {
+    async fn block_by_hash(&mut self, block_hash: Self::BlockHash) -> anyhow::Result<Self::Block> {
         let url = self.raw_block_by_hash_url(&block_hash);
-
         let client = self.client.clone();
-        let block = async move {
-            let block =
-                bitcoin_http_request_for_hex_encoded_object::<Self::Block>(url, client).await?;
-            tracing::debug!(
-                "Fetched block {} with {} transactions from bitcoind",
-                block_hash,
-                block.txdata.len()
-            );
+        let block = bitcoin_http_request_for_hex_encoded_object::<Self::Block>(url, client).await?;
 
-            Ok(block)
-        }
-        .boxed()
-        .compat();
+        tracing::debug!(
+            "Fetched block {} with {} transactions from bitcoind",
+            block_hash,
+            block.txdata.len()
+        );
 
-        Box::new(block)
+        Ok(block)
     }
 }
