+++ cnd/src/btsieve/bitcoin/cache.rs
@@ -2,8 +2,6 @@ use crate::btsieve::{BlockByHash, LatestBlock};
 use async_trait::async_trait;
 use bitcoin::{util::hash::BitcoinHash, Block, BlockHash as Hash, BlockHash};
 use derivative::Derivative;
-use futures::Future;
-use futures_core::{compat::Future01CompatExt, future::TryFutureExt};
 use lru::LruCache;
 use std::sync::Arc;
 use tokio::sync::Mutex;
@@ -50,4 +48,31 @@ where
     }
 }
 
-impl_block_by_hash!();
+#[async_trait]
+impl<C> BlockByHash for Cache<C>
+where
+    C: BlockByHash<Block = Block, BlockHash = Hash> + Clone,
+{
+    type Block = Block;
+    type BlockHash = BlockHash;
+
+    async fn block_by_hash(&mut self, block_hash: Self::BlockHash) -> anyhow::Result<Self::Block> {
+        let mut connector = self.connector.clone();
+        let cache = Arc::clone(&self.block_cache);
+
+        if let Some(block) = cache.lock().await.get(&block_hash) {
+            tracing::trace!("Found block in cache: {:x}", block_hash);
+            return Ok(block.clone());
+        }
+
+        let block = connector.block_by_hash(block_hash.clone()).await?;
+        tracing::trace!("Fetched block from connector: {:x}", block_hash);
+
+        // We dropped the lock so at this stage the block may have been inserted by
+        // another thread, no worries, inserting the same block twice does not hurt.
+        let mut guard = cache.lock().await;
+        guard.put(block_hash, block.clone());
+
+        Ok(block)
+    }
+}
