+++ cnd/tests/bitcoin_helper/connector_mock.rs
@@ -1,7 +1,7 @@
+use anyhow::Context;
 use async_trait::async_trait;
 use bitcoin::{hashes::sha256d, util::hash::BitcoinHash, BlockHash};
 use cnd::btsieve::{BlockByHash, LatestBlock};
-use futures::{future::IntoFuture, Future};
 use std::{
     collections::HashMap,
     time::{Duration, Instant},
@@ -59,21 +59,16 @@ impl LatestBlock for BitcoinConnectorMock {
     }
 }
 
+#[async_trait]
 impl BlockByHash for BitcoinConnectorMock {
     type Block = bitcoin::Block;
     type BlockHash = bitcoin::BlockHash;
 
-    fn block_by_hash(
-        &self,
-        block_hash: Self::BlockHash,
-    ) -> Box<dyn Future<Item = Self::Block, Error = anyhow::Error> + Send + 'static> {
-        Box::new(
-            self.all_blocks
-                .get(&block_hash)
-                .cloned()
-                .ok_or_else(|| anyhow::Error::from(Error::UnknownHash(block_hash)))
-                .into_future(),
-        )
+    async fn block_by_hash(&mut self, block_hash: Self::BlockHash) -> anyhow::Result<Self::Block> {
+        self.all_blocks
+            .get(&block_hash)
+            .cloned()
+            .with_context(|| format!("could not find block with hash {}", block_hash))
     }
 }
 
@@ -81,6 +76,4 @@ impl BlockByHash for BitcoinConnectorMock {
 pub enum Error {
     #[error("ran out of blocks in chain")]
     NoMoreBlocks,
-    #[error("could not find block with hash {0}")]
-    UnknownHash(BlockHash),
 }
